Use README as description; move read function to las.py
